Major code refactor.
