+++ packages/peregrine/lib/tapableHooks/index.js
@@ -1,0 +1,5 @@
+import { syncOnSignOut } from './myAccountHooks';
+
+export default {
+    syncOnSignOut
+};
